RestroStats_backend/requirements.txt
# alembic==1.14.1
# celery==5.4.0
# easyocr==1.7.2
Flask==3.1.0
Flask_Bcrypt==1.0.1
Flask_Cors==5.0.0
Flask_JWT_Extended==4.7.1
Flask_Migrate==4.1.0
flask_sqlalchemy==3.1.1
# matplotlib==3.10.1
numpy==2.2.5
# opencv_contrib_python==4.11.0.86
# opencv_python==4.11.0.86
# opencv_python_headless==4.11.0.86
pandas==2.2.3
# pyspellchecker==0.8.2
# pytesseract==0.3.13
# scikit_learn==1.6.1
# seaborn==0.13.2
SQLAlchemy==2.0.37
# pymysql==1.1.1
PyMySQL==1.1.1
cryptography==44.0.3